pyproject.toml
[tool.poetry]
name = "iso2dcat"
version = "0.1.0"
description = ""
authors = ["volker <volker.jaenisch@inqbus.de>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
OWSLib = "^0.25.0"
rdflib = "^6.1.1"
"zope.interface" = "^5.4.0"
"zope.component" = "^5.0.1"
colorlog = "^6.6.0"
dynaconf = "^3.1.7"
joblib = "^1.1.0"
prettytable = "^3.1.1"
more-itertools = "^8.12.0"
pyrdf4j = { path = "../pyrdf4j", develop = true }
pkan_config = {path="../pkan_config", develop=true}
lxml = "*"
pysolr = "*"
progressbar2 = "^4.0.0"
simplejson = "^3.17.6"


[tool.poetry.dev-dependencies]
coverage = "^6.3.2"
unittest2 = "^1.1.0"
pytest = "^7.1.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
